User manadgement

+++ server/src/middlewares/slots/deleteSlot.js
@@ -1,14 +1,13 @@
 import * as Slots from "../../model/Slots";
 
-export default (slotName) => {
+export default slotName => {
   const Slot = Slots[slotName];
   return async (req, res, next) => {
     try {
-      await Slot.deleteOne({ name: req.body.slot })
+      await Slot.deleteOne({ userId: req.userId, name: req.body.slot });
       return next();
-    }
-    catch (err) {
+    } catch (err) {
       return next(err);
     }
-  }
-}
+  };
+};
